+++ ft_itoa_base_capx.c
@@ -6,17 +6,17 @@
 /*   By: sehkang <sehkang@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/08 16:35:33 by sehkang           #+#    #+#             */
-/*   Updated: 2021/06/08 15:29:31 by sehkang          ###   ########.fr       */
+/*   Updated: 2021/06/09 10:25:41 by sehkang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-char		*ft_itoa_base_capx(ssize_t n)
+char		*ft_itoa_base_capx(long long n)
 {
-	size_t	size;
+	int		size;
 	char	*ret_ptr;
-	size_t	i;
+	int		i;
 	char	*base;
 
 	base = "0123456789ABCDEF";
